+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 -   Added craftcms/phpstan and run for pre-craft-4 cleanup
 -   Created `OrderValidator` service class and moved relevant code into it.
 -   Created `OrderMetadata` service class and moved relevant code into it.
+-   Move shipping request creation logic to model.
 
 ## 1.2.0 - 2023-11-06